requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
PyMySQL==1.0.2
Werkzeug==2.2.3
python-dotenv==0.21.1
gunicorn==20.1.0